pyproject.toml
[tool.poetry]
name = "scaffoldunified"
version = "0.1.0"
description = ""
authors = ["lhchavez <lhchavez@lhchavez.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
flask = "^3.0"
python-fasthtml = "^0.4.5"
gunicorn = "^23.0.0"
streamlit = "^1.12.1"
tenacity = "~8.5.0"
matplotlib = "^3.9.2"
replit = "^4.0.0"
pandas = "^2.2.2"
replit-object-storage = "^1.0.2"
inotify = "^0.2.10"
numpy = "~1.26.4"
openai = "^1.42.0"
trafilatura = "^1.12.1"
requests = "^2.32.3"
flask-sqlalchemy = "^3.1.1"
flask-jwt-extended = "^4.7.1"
flask-socketio = "^5.5.1"
flask-cors = "^6.0.1"
flask-migrate = "^4.1.0"
python-dotenv = "^1.2.1"
werkzeug = "^3.1.3"
psycopg2-binary = "^2.9.11"
python-dateutil = "^2.9.0.post0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
